refactored shows list and added location field
