+++ browser/src/stylesheets/artist.scss
@@ -105,24 +105,55 @@
 
 .recommended-shows {
   margin: 15px 0;
-  > h3 {
-    margin: 0;
-    font-size: 18px;
-  }
-  > hr {
-    margin: 5px 0 15px;
-  }
-  @media screen and (max-width: $screen-md) {
+  > .recommended-shows-header {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: flex-end;
     > h3 {
-      font-size: 16px;
+      margin: 0;
+      font-size: 18px;
     }
-  }
-  @media screen and (max-width: $screen-sm) {
-    > h3 {
-      font-size: 14px;
+    > .shows-location {
+      display: flex;
+      > input {
+        font-size: 16px;
+        width: 200px;
+      }
+      > button {
+        margin-left: 5px;
+      }
+      @media screen and (max-width: $screen-md) {
+        > input {
+          width: 150px;
+        }
+      }
+      @media screen and (max-width: $screen-sm) {
+        flex-direction: column;
+        align-items: center;
+        > input, > button {
+           width: 120px;
+        }
+        > button {
+          margin: 5px 0 0;
+        }
+      }
     }
     > hr {
-      margin-bottom: 10px;
+      margin: 5px 0 15px;
+    }
+    @media screen and (max-width: $screen-md) {
+      > h3 {
+        font-size: 16px;
+      }
+    }
+    @media screen and (max-width: $screen-sm) {
+      > h3 {
+        font-size: 14px;
+      }
+      > hr {
+        margin-bottom: 10px;
+      }
     }
   }
 }
