+++ browser/src/stylesheets/search.scss
@@ -5,20 +5,6 @@
 
 .artist-search {
   width: 100%;
-  border: none;
-  border-bottom: 2px solid $text-dark;
-  transition: border-bottom .3s linear;
-  font-size: 22px;
-  &:focus {
-    outline: none;
-    border-bottom-color: #28A3E1;
-  }
-  @media screen and (max-width: $screen-md) {
-    font-size: 18px;
-  }
-  @media screen and (max-width: $screen-sm) {
-    font-size: 14px;
-  }
 }
 
 .empty-search {
